APIs/amazonaws.com/acm-pca/2017-08-22/python-aiohttp-client/tests/test_default_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.create_certificate_authority_audit_report_request import CreateCertificateAuthorityAuditReportRequest
from openapi_server.models.create_certificate_authority_audit_report_response import CreateCertificateAuthorityAuditReportResponse
from openapi_server.models.create_certificate_authority_request import CreateCertificateAuthorityRequest
from openapi_server.models.create_certificate_authority_response import CreateCertificateAuthorityResponse
from openapi_server.models.create_permission_request import CreatePermissionRequest
from openapi_server.models.delete_certificate_authority_request import DeleteCertificateAuthorityRequest
from openapi_server.models.delete_permission_request import DeletePermissionRequest
from openapi_server.models.delete_policy_request import DeletePolicyRequest
from openapi_server.models.describe_certificate_authority_audit_report_request import DescribeCertificateAuthorityAuditReportRequest
from openapi_server.models.describe_certificate_authority_audit_report_response import DescribeCertificateAuthorityAuditReportResponse
from openapi_server.models.describe_certificate_authority_request import DescribeCertificateAuthorityRequest
from openapi_server.models.describe_certificate_authority_response import DescribeCertificateAuthorityResponse
from openapi_server.models.get_certificate_authority_certificate_request import GetCertificateAuthorityCertificateRequest
from openapi_server.models.get_certificate_authority_certificate_response import GetCertificateAuthorityCertificateResponse
from openapi_server.models.get_certificate_authority_csr_request import GetCertificateAuthorityCsrRequest
from openapi_server.models.get_certificate_authority_csr_response import GetCertificateAuthorityCsrResponse
from openapi_server.models.get_certificate_request import GetCertificateRequest
from openapi_server.models.get_certificate_response import GetCertificateResponse
from openapi_server.models.get_policy_request import GetPolicyRequest
from openapi_server.models.get_policy_response import GetPolicyResponse
from openapi_server.models.import_certificate_authority_certificate_request import ImportCertificateAuthorityCertificateRequest
from openapi_server.models.issue_certificate_request import IssueCertificateRequest
from openapi_server.models.issue_certificate_response import IssueCertificateResponse
from openapi_server.models.list_certificate_authorities_request import ListCertificateAuthoritiesRequest
from openapi_server.models.list_certificate_authorities_response import ListCertificateAuthoritiesResponse
from openapi_server.models.list_permissions_request import ListPermissionsRequest
from openapi_server.models.list_permissions_response import ListPermissionsResponse
from openapi_server.models.list_tags_request import ListTagsRequest
from openapi_server.models.list_tags_response import ListTagsResponse
from openapi_server.models.put_policy_request import PutPolicyRequest
from openapi_server.models.restore_certificate_authority_request import RestoreCertificateAuthorityRequest
from openapi_server.models.revoke_certificate_request import RevokeCertificateRequest
from openapi_server.models.tag_certificate_authority_request import TagCertificateAuthorityRequest
from openapi_server.models.untag_certificate_authority_request import UntagCertificateAuthorityRequest
from openapi_server.models.update_certificate_authority_request import UpdateCertificateAuthorityRequest


pytestmark = pytest.mark.asyncio

async def test_create_certificate_authority(client):
    """Test case for create_certificate_authority

    
    """
    body = {"IdempotencyToken":"","CertificateAuthorityConfiguration":{"CsrExtensions":{"KeyUsage":{"KeyEncipherment":"","DataEncipherment":"","DigitalSignature":"","KeyCertSign":"","DecipherOnly":"","KeyAgreement":"","NonRepudiation":"","CRLSign":"","EncipherOnly":""},"SubjectInformationAccess":""},"SigningAlgorithm":"","Subject":{"Organization":"","OrganizationalUnit":"","Locality":"","Title":"","GivenName":"","GenerationQualifier":"","Initials":"","CustomAttributes":"","SerialNumber":"","State":"","Country":"","Surname":"","DistinguishedNameQualifier":"","CommonName":"","Pseudonym":""},"KeyAlgorithm":""},"RevocationConfiguration":{"OcspConfiguration":{"OcspCustomCname":"","Enabled":""},"CrlConfiguration":{"CustomCname":"","S3ObjectAcl":"","ExpirationInDays":"","Enabled":"","S3BucketName":""}},"CertificateAuthorityType":"","UsageMode":"","KeyStorageSecurityStandard":"","Tags":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.CreateCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_create_certificate_authority_audit_report(client):
    """Test case for create_certificate_authority_audit_report

    
    """
    body = {"CertificateAuthorityArn":"","AuditReportResponseFormat":"","S3BucketName":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.CreateCertificateAuthorityAuditReport',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_create_permission(client):
    """Test case for create_permission

    
    """
    body = {"CertificateAuthorityArn":"","Actions":"","SourceAccount":"","Principal":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.CreatePermission',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_certificate_authority(client):
    """Test case for delete_certificate_authority

    
    """
    body = {"CertificateAuthorityArn":"","PermanentDeletionTimeInDays":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.DeleteCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_permission(client):
    """Test case for delete_permission

    
    """
    body = {"CertificateAuthorityArn":"","SourceAccount":"","Principal":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.DeletePermission',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_policy(client):
    """Test case for delete_policy

    
    """
    body = {"ResourceArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.DeletePolicy',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_describe_certificate_authority(client):
    """Test case for describe_certificate_authority

    
    """
    body = {"CertificateAuthorityArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.DescribeCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_describe_certificate_authority_audit_report(client):
    """Test case for describe_certificate_authority_audit_report

    
    """
    body = {"CertificateAuthorityArn":"","AuditReportId":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.DescribeCertificateAuthorityAuditReport',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_certificate(client):
    """Test case for get_certificate

    
    """
    body = {"CertificateAuthorityArn":"","CertificateArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.GetCertificate',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_certificate_authority_certificate(client):
    """Test case for get_certificate_authority_certificate

    
    """
    body = {"CertificateAuthorityArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.GetCertificateAuthorityCertificate',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_certificate_authority_csr(client):
    """Test case for get_certificate_authority_csr

    
    """
    body = {"CertificateAuthorityArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.GetCertificateAuthorityCsr',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_policy(client):
    """Test case for get_policy

    
    """
    body = {"ResourceArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.GetPolicy',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_import_certificate_authority_certificate(client):
    """Test case for import_certificate_authority_certificate

    
    """
    body = {"CertificateAuthorityArn":"","CertificateChain":"","Certificate":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.ImportCertificateAuthorityCertificate',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_issue_certificate(client):
    """Test case for issue_certificate

    
    """
    body = {"TemplateArn":"","CertificateAuthorityArn":"","Validity":{"Type":"","Value":""},"IdempotencyToken":"","Csr":"","SigningAlgorithm":"","ApiPassthrough":{"Extensions":{"CustomExtensions":"","CertificatePolicies":"","KeyUsage":{"KeyEncipherment":"","DataEncipherment":"","DigitalSignature":"","KeyCertSign":"","DecipherOnly":"","KeyAgreement":"","NonRepudiation":"","CRLSign":"","EncipherOnly":""},"SubjectAlternativeNames":"","ExtendedKeyUsage":""},"Subject":{"Organization":"","OrganizationalUnit":"","Locality":"","Title":"","GivenName":"","GenerationQualifier":"","Initials":"","CustomAttributes":"","SerialNumber":"","State":"","Country":"","Surname":"","DistinguishedNameQualifier":"","CommonName":"","Pseudonym":""}},"ValidityNotBefore":{"Type":"","Value":""}}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.IssueCertificate',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_certificate_authorities(client):
    """Test case for list_certificate_authorities

    
    """
    body = {"NextToken":"","MaxResults":"","ResourceOwner":""}
    params = [('MaxResults', 'max_results_example'),
                    ('NextToken', 'next_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.ListCertificateAuthorities',
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_permissions(client):
    """Test case for list_permissions

    
    """
    body = {"CertificateAuthorityArn":"","NextToken":"","MaxResults":""}
    params = [('MaxResults', 'max_results_example'),
                    ('NextToken', 'next_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.ListPermissions',
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_tags(client):
    """Test case for list_tags

    
    """
    body = {"CertificateAuthorityArn":"","NextToken":"","MaxResults":""}
    params = [('MaxResults', 'max_results_example'),
                    ('NextToken', 'next_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.ListTags',
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_put_policy(client):
    """Test case for put_policy

    
    """
    body = {"Policy":"","ResourceArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.PutPolicy',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_restore_certificate_authority(client):
    """Test case for restore_certificate_authority

    
    """
    body = {"CertificateAuthorityArn":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.RestoreCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_revoke_certificate(client):
    """Test case for revoke_certificate

    
    """
    body = {"RevocationReason":"","CertificateAuthorityArn":"","CertificateSerial":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.RevokeCertificate',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tag_certificate_authority(client):
    """Test case for tag_certificate_authority

    
    """
    body = {"CertificateAuthorityArn":"","Tags":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.TagCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_untag_certificate_authority(client):
    """Test case for untag_certificate_authority

    
    """
    body = {"CertificateAuthorityArn":"","Tags":""}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.UntagCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_update_certificate_authority(client):
    """Test case for update_certificate_authority

    
    """
    body = {"Status":"","CertificateAuthorityArn":"","RevocationConfiguration":{"OcspConfiguration":{"OcspCustomCname":"","Enabled":""},"CrlConfiguration":{"CustomCname":"","S3ObjectAcl":"","ExpirationInDays":"","Enabled":"","S3BucketName":""}}}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'x_amz_content_sha256': 'x_amz_content_sha256_example',
        'x_amz_date': 'x_amz_date_example',
        'x_amz_algorithm': 'x_amz_algorithm_example',
        'x_amz_credential': 'x_amz_credential_example',
        'x_amz_security_token': 'x_amz_security_token_example',
        'x_amz_signature': 'x_amz_signature_example',
        'x_amz_signed_headers': 'x_amz_signed_headers_example',
        'x_amz_target': 'x_amz_target_example',
        'hmac': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/#X-Amz-Target=ACMPrivateCA.UpdateCertificateAuthority',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

